Fix BEATs package imports and the AMP crash in the weak loss
Both bugs only appear on a GPU Colab run, which local CPU testing missed.

- third_party/beats: upstream uses flat imports (`from backbone import ...`)
  that only resolve when that directory is on sys.path, so importing it as a
  package failed with "No module named 'backbone'" and silently fell back to
  mel-only. Rewritten to relative imports; nothing else changed.
- losses: F.binary_cross_entropy is rejected under autocast. The attention-pooled
  clip probability is a weighted average of probabilities and has no pre-pooling
  logit, so the _with_logits form does not apply; the weak BCE is now written out
  in float32 by hand.
- fdy_crnn: pool in float32 - the 1e-7 clamp underflows in fp16.
- dataset: draw with replacement when a tier is smaller than its batch quota,
  so batches keep their shape on small/early dataset batches.

Verified: BEATs loads (90.31M frozen / 8.97M trainable), emits 6.2 fps features
that fuse and interpolate to the output grid, and trains + predicts end to end on
real Vaani audio. New tests/test_amp_loss.py covers the autocast regression.

+++ src/models/fdy_crnn.py
@@ -127,7 +127,9 @@ class AttentionPool(nn.Module):
         a = self.att(h)
         if frame_valid is not None:
             a = a.masked_fill(frame_valid[..., None] < 0.5, torch.finfo(a.dtype).min)
-        a = torch.softmax(a, dim=1)
-        strong = torch.sigmoid(strong_logit)
+        # Pool in float32: under AMP the softmax-weighted average and the 1e-7
+        # clamp both underflow in fp16 (smallest normal half is ~6e-5).
+        a = torch.softmax(a.float(), dim=1)
+        strong = torch.sigmoid(strong_logit.float())
         clip = (strong * a).sum(dim=1).clamp(1e-7, 1 - 1e-7)
         return strong_logit, clip
